requirements.txt
# =============================================================================
# 智能辩论系统 (Argue Agents)
# 依赖配置文件
# =============================================================================
# 作者: Jiangsheng Yu
# =============================================================================

# 核心依赖 - 大语言模型
transformers>=4.37.0
torch>=2.0.0
accelerate>=0.25.0

# 模型支持
sentencepiece
protobuf

# RAG (检索增强生成)
faiss-cpu
sentence-transformers

# 网络搜索
googlesearch-python
requests
beautifulsoup4

# 终端美化
rich
